Cognifyz Internship Tasks
Level 1
Task 1 : String Reversal
Create a Python function that takes a string as input and returns the reverse of that string. For example, if the input is "hello," the function should return "olleh."

Task 2 : Temperature Conversion
Create a Python program that converts temperatures between Celsius and Fahrenheit. Prompt the user to enter a temperature value and the unit of measurement, and then display the converted temperature.

Task 3 : Email Validator
Develop a Python function that validates whether a given string is a valid email address. Implement checks for the format, including the presence of an "@" symbol and a domain name.

Task 4 : Calculator Program
Create a Python program that acts as a basic calculator. It should prompt the user to enter two numbers and an operator (+, -, *, /, %), and then display the result of the operation.

Task 5 : Palindrome Checker
Write a Python function that checks whether a given string is a palindrome. A palindrome is a word, phrase, or sequence that reads the same backward as forward (e.g., "madam" or "racecar").

Level 2
Task 1 : Guessing Game
Write a Python program that generates a random number between 1 and 100. The user should then try to guess the number .The program should provide hints such as "too high" or "too low" until the correct number is guessed.

Task 2 : Number Guesser
Create a number guessing game where the program generates a random number between a specified range, and the user tries to guess it. Provide feedback to the user if their guess is too high or too low.

Task 3 : Password Strength Checker
Create a Python function that evaluates the strength of a password entered by the user. Implement checks for factors such as length, presence of uppercase and lowercase letters, digits, and special characters.

Task 4 : Fibonacci Sequence
Write a Python function that generates the Fibonacci sequence up to a given number of terms. The function should take an integer input from the user and display the Fibonacci sequence up to that number of terms.

Task 5 : File Manipulation
Write a Python program that reads a text file and counts the occurrences of each word in the file. Display the results in alphabetical order along with their respective counts.

Level 3
Task: Build a Web Scraper
Develop a web scraper that extracts specific data from websites using libraries like BeautifulSoup or Scrapy. This task will improve their knowledge of web scraping techniques and handling HTML/XML data.